3.0版本开发

+++ src/main/java/com/mtons/mblog/modules/service/impl/PostSearchServiceImpl.java
@@ -57,7 +57,7 @@ public class PostSearchServiceImpl implements PostSearchService {
                 .matching(q).createQuery();
 
         org.hibernate.search.jpa.FullTextQuery query = fullTextSession.createFullTextQuery(luceneQuery);
-        query.setFirstResult(pageable.getOffset());
+        query.setFirstResult((int) pageable.getOffset());
         query.setMaxResults(pageable.getPageSize());
 
         StandardAnalyzer standardAnalyzer = new StandardAnalyzer();
@@ -100,7 +100,7 @@ public class PostSearchServiceImpl implements PostSearchService {
         luceneQuery = term.createQuery();
 
         org.hibernate.search.jpa.FullTextQuery query = fullTextSession.createFullTextQuery(luceneQuery);
-        query.setFirstResult(pageable.getOffset());
+        query.setFirstResult((int) pageable.getOffset());
         query.setMaxResults(pageable.getPageSize());
 
         Sort sort = new Sort(new SortField("id", SortField.Type.LONG, true));
